Apply review follow-ups for service API boundaries and defaults
