+++ reticulum_telemetry_hub/reticulum_server/delivery_defaults.py
@@ -1,0 +1,9 @@
+"""Shared outbound delivery default values."""
+
+DEFAULT_OUTBOUND_QUEUE_SIZE = 64
+DEFAULT_OUTBOUND_WORKERS = 4
+DEFAULT_OUTBOUND_SEND_TIMEOUT = 5.0
+DEFAULT_OUTBOUND_DELIVERY_RECEIPT_TIMEOUT = 30.0
+DEFAULT_OUTBOUND_BACKOFF = 0.5
+DEFAULT_OUTBOUND_MAX_ATTEMPTS = 3
+DEFAULT_OUTBOUND_FANOUT_SOFT_MAX_RECIPIENTS = 256
